new dirichlet

+++ Nethermind.Dirichlet.Numerics/UInt256.cs
@@ -164,7 +164,7 @@ namespace Nethermind.Dirichlet.Numerics
                 throw new FormatException();
             return c;
         }
-        
+
         public static UInt256 Parse(ReadOnlySpan<char> value, NumberStyles numberStyles)
         {
             if (!TryParse(value, numberStyles, CultureInfo.InvariantCulture, out UInt256 c))
@@ -176,7 +176,7 @@ namespace Nethermind.Dirichlet.Numerics
         {
             return TryParse(value, NumberStyles.Integer, CultureInfo.InvariantCulture, out result);
         }
-        
+
         public static bool TryParse(ReadOnlySpan<char> value, out UInt256 result)
         {
             return TryParse(value, NumberStyles.Integer, CultureInfo.InvariantCulture, out result);
@@ -186,7 +186,7 @@ namespace Nethermind.Dirichlet.Numerics
         {
             return TryParse(value.AsSpan(), style, provider, out result);
         }
-        
+
         public static bool TryParse(ReadOnlySpan<char> value, NumberStyles style, IFormatProvider provider, out UInt256 result)
         {
             BigInteger a;
@@ -199,12 +199,12 @@ namespace Nethermind.Dirichlet.Numerics
                 bigParsedProperly = BigInteger.TryParse(fixedHexValue, style, provider, out a);
             }
             else
-            { 
-                Span<char> fixedHexValue =  stackalloc char[value.Length];
+            {
+                Span<char> fixedHexValue = stackalloc char[value.Length];
                 value.CopyTo(fixedHexValue);
                 bigParsedProperly = BigInteger.TryParse(fixedHexValue, style, provider, out a);
             }
-            
+
             if (!bigParsedProperly)
             {
                 result = Zero;
@@ -294,15 +294,15 @@ namespace Nethermind.Dirichlet.Numerics
             // swap adjacent 8-bit blocks
             return ((x & 0xFF00FF00FF00FF00) >> 8) | ((x & 0x00FF00FF00FF00FF) << 8);
         }
-        
+
         public static ulong SwapBytes2(ulong value)
         {
-            return BinaryPrimitives.ReverseEndianness(value);
+            return (ulong) BinaryPrimitives.ReverseEndianness((long) value);
         }
-        
+
         public static ulong SwapBytes3(ulong value)
         {
-            return (ulong)IPAddress.HostToNetworkOrder((long)value);
+            return (ulong) IPAddress.HostToNetworkOrder((long) value);
         }
 
         public static void CreateFromBigEndian(out UInt256 c, Span<byte> span)
@@ -1857,7 +1857,7 @@ namespace Nethermind.Dirichlet.Numerics
         public static uint Remainder(ref UInt256 u, uint v)
         {
             if (u.s1 == 0)
-                return (uint)(u.s0 % v);
+                return (uint) (u.s0 % v);
             throw new NotImplementedException();
 //            if (u.s1 <= uint.MaxValue)
 //                return Remainder96(ref u, v);
